+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to the "glsl-canvas" library will be documented in this file
 
 ---
 
+## [0.1.1] - 2019-02-06
+### Added
+- Texture error event.
+### Fixed
+- Video texture.
+
+---
+
 ## [0.1.0] - 2019-02-02
 ### Added
 - Initial release of glsl-canvas lib.